% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseRelatedFunctions.R
\name{check.database}
\alias{check.database}
\title{Check the data in the reference database}
\usage{
check.database(path)
}
\arguments{
\item{path}{character. full path to the database file (*.sqlite)}
}
\value{
This function returns a data table including the summary of the specified database. This is neccessary to check the consistency and validity of an unknown or new database file.
}
\description{
This function displays the summary of the database, including how many tables are in the database file, number of data rows for each data table and the first row of each table.
}
\note{
First column include the names of the tables in database

Second column is the number of rows in each table

Next columns are the first row of each table
}
\examples{
check.database(system.file("extdata", "sample_db.sqlite", package = "GWASinspector"))
}
