% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_GWASs.R
\name{compare.GWASs}
\alias{compare.GWASs}
\title{Compare GWAS result files}
\usage{
compare.GWASs(input.file.list, output.path)
}
\arguments{
\item{input.file.list}{list, full path of the RDS object files. Note that inspect() only produces such files if the object_file parameter is set to TRUE in the configuration file.}

\item{output.path}{character, full path to the folder where output files should be saved.}
}
\value{
Key metrics report of previously inspected files are generated and saved in the specified output folder.
}
\description{
This function compares the key metrics of previously inspected files.
This allows the user to check that the results of these studies are comparable (important when running a meta-analysis) and that there are no significant anomalies.
}
