% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_inspector.R
\name{demo_inspector}
\alias{demo_inspector}
\title{Runs an example QC}
\usage{
demo_inspector(result.dir)
}
\arguments{
\item{result.dir}{character. Path to the output folder for saving QC result files}
}
\value{
QC reports from running the algorithm on a sample GWAS file are generated and saved in the specified folder.
}
\description{
This function runs the QC algorithm on a fabricated GWAS result file.
}
