% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfig.R
\name{get.config}
\alias{get.config}
\title{Copies the template configuration file to the local machine}
\usage{
get.config(dir.path)
}
\arguments{
\item{dir.path}{Path to the folder for saving a sample configuration file.}
}
\value{
Copies a sample configuration file (config.ini) in the specified folder.
}
\description{
This templates should be edited and then used for setting up and running the QC pipeline.
Default filename is \strong{config.ini}.
}
\examples{
get.config(tempdir())

}
