% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeaderTranslation.R
\name{get_headerTranslation}
\alias{get_headerTranslation}
\title{Copies the template header translation table to the local machine}
\usage{
get_headerTranslation(dir.path)
}
\arguments{
\item{dir.path}{Path to the folder for saving a header-translation table file.}
}
\value{
Copies a sample header-translation table in the specified folder.
}
\description{
This template file is used to translate a dataset`s column names (the header) into the standard names used by GWASinspector.
The file contains a two-column table, with the left column containing the standard column-names and the right the alternatives.
Both the standard and alternative columns must be fully capitalized. This is a text file which includes most common variable/header names and can be edited according to user specifications.
The default filename is \strong{alt_headers.txt}.
}
