% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_GWAS.R
\name{inspect}
\alias{inspect}
\title{Runs the QC algorithm}
\usage{
inspect(config.file = NULL, user.verification = FALSE,
  test.run = FALSE)
}
\arguments{
\item{config.file}{character. Path to a configuration (.ini) file to configurate the QC. For a sample configuration file, see get.config(). For details on how to edit the config file, see the tutorial.}

\item{user.verification}{logical. If TRUE, the algorithm will pause and ask the user to verify that it has selected the correct input files for QC. Default vaule is FALSE.}

\item{test.run}{logical. if TRUE, only the first 1000 lines of each data file are loaded and analysed, and no plots or final dataset is produced.}
}
\value{
QC reports from running the algorithm on a single or a series of GWAS result files are generated and saved.
}
\description{
This is the main function for running the algorithm on GWAS result files.
}
