% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileFunctions.R
\name{reformat.columns}
\alias{reformat.columns}
\title{Unifies column separators in all data rows}
\usage{
reformat.columns(input.file, output.file = NULL, sep = "\\t")
}
\arguments{
\item{input.file}{path to the input file}

\item{output.file}{path of the outfile path; if not defined , a tag is added to input file}

\item{sep}{what character to be used add separator; tab character is default}
}
\value{
A new text file is generated from the old one with a consistent column separator character.
}
\description{
It is common that column separator is different between the header and the rest of the file. This can be problematic for reading the file correctly and should be treated before loading the data.
This function is based on Rtools and *awk* function to read and save a new file with equal column separators between all rows of data.
}
