% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_inspector.R
\name{result.inspector}
\alias{result.inspector}
\title{Display the reports of running the Inspector algorithm}
\usage{
result.inspector(inspector)
}
\arguments{
\item{inspector}{An instance of \linkS4class{Inspector} class. Check \code{\link{setup.inspector}} for more details.}
}
\value{
A data.table containing a brief report about the results.
}
\description{
This function displays a brief report about the results of running the Inspector algorithm on a list of GWAS result files.
The full report including plots, cleaned files and summary statistics are generated and saved in the output folder during the algorithm run.
}
