% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inspector.R
\name{run.inspector}
\alias{run.inspector}
\title{Runs the QC pipeline on a set of GWAS result files}
\usage{
run.inspector(inspector, verbose = TRUE, test.run = FALSE)
}
\arguments{
\item{inspector}{An instance of \linkS4class{Inspector} class. Check \code{\link{setup.inspector}} for more details.}

\item{verbose}{logical. If FALSE, no messages will show up in the terminal and are only saved in the log file.}

\item{test.run}{logical. If TRUE, only the first 1000 lines of each data file are loaded and analyzed;
plots and saving the cleaned output dataset are skipped. Default value is FALSE.}
}
\value{
Reports from running the algorithm on a single or a series of GWAS result files are generated and saved.
}
\description{
This is the main function of the package for running the QC algorithm on a set of GWAS result files.
It requires an object of class \linkS4class{Inspector} which should be created by \code{\link{setup.inspector}}.
Check the package vignette and tutorial for more details on this topic.
}
\examples{
\dontrun{
# save a sample configuration file and edit it as required.
config.file <- get.config(tempdir())

# import the QC-configuration file into R.
job <- setup.inspector(config.file)

# check the generated object.
job

# run the generated object.
# This may take up to a couple of hours, depending on your PC and the size/number of files.
job <- run.inspector(job)

# view a brief summary about the result. report files and plots are already saved.
result.inspector(job)
}

}
