% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inspector.R
\name{run_inspector}
\alias{run_inspector}
\title{Runs the QC pipeline on a set of GWAS result files}
\usage{
run_inspector(inspector, verbose = TRUE, test.run = FALSE)
}
\arguments{
\item{inspector}{An instance of \linkS4class{Inspector} class. Check \code{\link{setup_inspector}} for more details.}

\item{verbose}{logical. If FALSE, no messages will show up in the terminal and are only saved in the log file.}

\item{test.run}{logical. If TRUE, only the first 1000 lines of each data file are loaded and analyzed;
plots and saving the cleaned output dataset are skipped. Default value is FALSE.}
}
\value{
Reports from running the algorithm on a single or a series of GWAS result files are generated and saved.
}
\description{
This is the main function of the package for running the QC algorithm on a set of GWAS result files.
It requires an object of class \linkS4class{Inspector} which should be created by \code{\link{setup_inspector}}.
Check the package vignette and tutorial for more details on this topic.
}
