% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definition.R
\docType{class}
\name{Study-class}
\alias{Study-class}
\alias{Study}
\title{An S4 class to represent an inspected GWAS study result file.}
\description{
This class is embedded in the \linkS4class{StudyList} class and should not be initiated separately.
}
\section{Slots}{

\describe{
\item{\code{File}}{A list representing GWAS result file specifications}

\item{\code{Counts}}{A list representing different variant counts from the GWAS result file.}

\item{\code{Correlations}}{A list representing different allele frequency and P-value correlations in the GWAS result file.}

\item{\code{Statistics}}{A list representing summary statistics from the GWAS result file.}

\item{\code{Successful_run}}{A logical value indicating whether the run was successful or not.}

\item{\code{starttime}}{The time that file inspection started.}

\item{\code{endtime}}{The time that file inspection ended.}
}}

