% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definition.R
\docType{class}
\name{StudyList-class}
\alias{StudyList-class}
\alias{StudyList}
\title{An S4 class to represent a list of inspected GWAS study result files.}
\description{
This class is embedded in the \linkS4class{Inspector} class and should not be initiated separately.
}
\section{Slots}{

\describe{
\item{\code{studyList}}{A list of GWAS study result files. Each member of this list is of class \code{\linkS4class{Study}}.}

\item{\code{studyCount}}{A numeric value indicating how many items of class \code{\linkS4class{Study}} are included.}
}}

