% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_check.R
\name{system.check}
\alias{system.check}
\title{Checks which required and optional packages are available}
\usage{
system.check()
}
\value{
System information and required functionalities for the QC algorythm are checked and reported as a data frame.
}
\description{
This functions checks the availability of all required and optional packages for GWASinspector.
The optional packages are not mandatory for running the algorithm; but will add useful functionalities.
}
\examples{
system.check()

}
