% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_check.R
\name{system.check}
\alias{system.check}
\title{Checks which required and optional packages are available}
\usage{
system.check()
}
\value{
System information and required functionalities for the QC algorithm are checked and reported as a data frame.
}
\description{
Checks if required and optional packages are installed on the system.
Although the optional packages do not contribute to the QC itself, having them available
will allow for Excel and HTML formatted log files, which are easier to read and interpret.
}
